+++ VisitorFramework/include/Visitor/Acyclic/VisitableAdapter.h
